fixed rect class

+++ include/Utils/CRect.h
@@ -4,11 +4,11 @@ namespace Utils {
     class CRect {
 
     private:
-        int x,y;
+        float x,y;
         int w,h;
     public:
         static CRect Zero();
-        CRect(int xa = 0, int ya = 0, int wa = 0, int ha = 0);
+        CRect(float xa = 0, float ya = 0, int wa = 0, int ha = 0);
 
         CRect& operator=(const CRect &r){
             if(this == &r) return *this;
@@ -32,11 +32,11 @@ namespace Utils {
         int Top() const ;
         int Bottom() const;
 
-        void setSize(int x, int y, int w, int h);
-        int getX() const;
-        int getY() const;
-        int getW() const;
-        int getH() const;
+        void setSize(float x, float y, int w, int h);
+        float getX() const;
+        float getY() const;
+        float getW() const;
+        float getH() const;
 
 
     };
