+++ src/Utils/CRect.cpp
@@ -6,7 +6,7 @@ Utils::CRect Utils::CRect::Zero()
 }
 
 
-Utils::CRect::CRect(int xa, int ya, int wa, int ha) : x(xa), y(ya), w(wa), h(ha){
+Utils::CRect::CRect(float xa, float ya, int wa, int ha) : x(xa), y(ya), w(wa), h(ha){
 
 }
 
@@ -32,22 +32,22 @@ bool Utils::CRect::HasIntersection(Utils::CRect b) const
 }
 
 INLINE_FUNCTION int Utils::CRect::Left() const{
-    return x;
+    return static_cast<int>(x);
 }
 
 INLINE_FUNCTION int Utils::CRect::Right() const{
-    return x + w;
+    return static_cast<int>(x + w);
 }
 
 INLINE_FUNCTION int Utils::CRect::Top() const {
-    return y;
+    return static_cast<int>(y);
 }
 
 INLINE_FUNCTION int Utils::CRect::Bottom() const {
-    return y + h;
+    return static_cast<int>(y + h);
 }
 
-void Utils::CRect::setSize(int x, int y, int w, int h)
+void Utils::CRect::setSize(float x, float y, int w, int h)
 {
     this->x = x;
     this->y = y;
@@ -56,22 +56,22 @@ void Utils::CRect::setSize(int x, int y, int w, int h)
     return;
 }
 
-int Utils::CRect::getX() const
+float Utils::CRect::getX() const
 {
     return x;
 }
 
-int Utils::CRect::getY() const
+float Utils::CRect::getY() const
 {
     return y;
 }
 
-int Utils::CRect::getW() const
+float Utils::CRect::getW() const
 {
     return w;
 }
 
-int Utils::CRect::getH() const
+float Utils::CRect::getH() const
 {
     return h;
 }
